requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
transformers
datasets
sentence-transformers
pandas
pydantic
nltk
rank_bm25
xgboost
